Create Grading-system.py
This Python program evaluates a student's score and assigns a grade from A to F using if-else statements. It validates the input to ensure it's a whole number between 0 and 100. If the score is valid, it prints the corresponding grade along with “YES”; otherwise, it displays an error message.

Name: Chidiebere Emmanuel Chidiebere 
Reg no: 2024704020